refactor: move WPCOM oauth-related code to separate file
