+++ includes/wizards/class-support-wizard.php
@@ -17,9 +17,6 @@ require_once NEWSPACK_ABSPATH . '/includes/wizards/class-wizard.php';
  * Interface for support for Newspack customers.
  */
 class Support_Wizard extends Wizard {
-	const NEWSPACK_WPCOM_ACCESS_TOKEN = '_newspack_wpcom_access_token';
-	const NEWSPACK_WPCOM_EXPIRES_IN   = '_newspack_wpcom_expires_in';
-
 	/**
 	 * The slug of this wizard.
 	 *
@@ -67,78 +64,6 @@ class Support_Wizard extends Wizard {
 				'permission_callback' => [ $this, 'api_permissions_check' ],
 			]
 		);
-
-		// Handle access token from WPCOM.
-		register_rest_route(
-			NEWSPACK_API_NAMESPACE,
-			'/wizard/newspack-support-wizard/wpcom_access_token',
-			[
-				'methods'             => \WP_REST_Server::CREATABLE,
-				'callback'            => [ $this, 'api_wpcom_access_token' ],
-				'permission_callback' => [ $this, 'api_permissions_check' ],
-			]
-		);
-
-		// Validate WPCOM access token.
-		register_rest_route(
-			NEWSPACK_API_NAMESPACE,
-			'/wizard/newspack-support-wizard/validate-access-token',
-			[
-				'methods'             => \WP_REST_Server::READABLE,
-				'callback'            => [ $this, 'api_wpcom_validate_access_token' ],
-				'permission_callback' => [ $this, 'api_permissions_check' ],
-			]
-		);
-	}
-
-	/**
-	 * Retrieve WPCOM access token.
-	 */
-	public static function get_wpcom_access_token() {
-		return get_user_meta( get_current_user_id(), self::NEWSPACK_WPCOM_ACCESS_TOKEN, true );
-	}
-
-	/**
-	 * Validate WPCOM credentials.
-	 */
-	public function api_wpcom_validate_access_token() {
-		$access_token = self::get_wpcom_access_token();
-		$client_id    = self::wpcom_client_id();
-		$response     = wp_safe_remote_get(
-			'https://public-api.wordpress.com/oauth2/token-info?' . http_build_query(
-				array(
-					'client_id' => $client_id,
-					'token'     => $access_token,
-				)
-			)
-		);
-		if ( is_wp_error( $response ) ) {
-			return $response;
-		}
-		if ( 200 !== $response['response']['code'] ) {
-			return new WP_Error( 'invalid_wpcom_token', __( 'Invalid WPCOM token.', 'newspack' ) );
-		}
-		return $response;
-	}
-
-	/**
-	 * Save WPCOM credentials.
-	 *
-	 * @param WP_REST_Request $request The request.
-	 */
-	public function api_wpcom_access_token( $request ) {
-		if ( isset( $request['access_token'], $request['expires_in'] ) ) {
-			$user_id = get_current_user_id();
-			update_user_meta( $user_id, self::NEWSPACK_WPCOM_ACCESS_TOKEN, sanitize_text_field( $request['access_token'] ) );
-			update_user_meta( $user_id, self::NEWSPACK_WPCOM_EXPIRES_IN, sanitize_text_field( $request['expires_in'] ) );
-			return \rest_ensure_response(
-				array(
-					'status' => 'saved',
-				)
-			);
-		} else {
-			return new WP_Error( 'missing_parameters', __( 'Missing parameters in request.', 'newspack' ) );
-		}
 	}
 
 	/**
@@ -155,7 +80,7 @@ class Support_Wizard extends Wizard {
 		}
 
 		try {
-			$wpcom_user_data = self::perform_wpcom_api_request( 'rest/v1.1/me' );
+			$wpcom_user_data = WPCOM_OAuth::perform_wpcom_api_request( 'rest/v1.1/me' );
 		} catch ( \Exception $e ) {
 			return new WP_Error(
 				'newspack_support_error',
@@ -224,58 +149,14 @@ class Support_Wizard extends Wizard {
 	}
 
 	/**
-	 * Get WPCOM access token.
-	 */
-	public static function get_access_token() {
-		$access_token = get_user_meta( get_current_user_id(), self::NEWSPACK_WPCOM_ACCESS_TOKEN, true );
-		if ( ! $access_token ) {
-			return new WP_Error(
-				'newspack_support_error',
-				__( 'Missing WPCOM access token.', 'newspack' )
-			);
-		}
-		return $access_token;
-	}
-
-	/**
-	 * Perform WPCOM API Request.
-	 *
-	 * @param string $endpoint endpoint.
-	 * @throws \Exception Error message.
-	 */
-	public static function perform_wpcom_api_request( $endpoint ) {
-		$access_token = self::get_access_token();
-		if ( is_wp_error( $access_token ) ) {
-			return $access_token;
-		}
-		$response = wp_safe_remote_get(
-			'https://public-api.wordpress.com/' . $endpoint,
-			array(
-				'headers' => [
-					'Authorization' => 'Bearer ' . $access_token,
-				],
-			)
-		);
-		if ( is_wp_error( $response ) ) {
-			throw new \Exception( $response->get_error_message() );
-		}
-		$response_body = json_decode( $response['body'] );
-		if ( $response['response']['code'] >= 300 ) {
-			throw new \Exception( $response['response']['message'] );
-		}
-		return $response_body;
-	}
-
-
-	/**
 	 * List support history.
 	 *
 	 * @param WP_REST_Request $request Request object.
 	 */
 	public static function api_support_history( $request ) {
 		try {
-			$user_data    = self::perform_wpcom_api_request( 'rest/v1.1/me' );
-			$support_data = self::perform_wpcom_api_request( 'wpcom/v2/support-history?email=' . $user_data->email );
+			$user_data    = WPCOM_OAuth::perform_wpcom_api_request( 'rest/v1.1/me' );
+			$support_data = WPCOM_OAuth::perform_wpcom_api_request( 'wpcom/v2/support-history?email=' . $user_data->email );
 			return $support_data->data;
 		} catch ( \Exception $e ) {
 			return new WP_Error(
@@ -330,9 +211,9 @@ class Support_Wizard extends Wizard {
 			true
 		);
 
-		$client_id    = self::wpcom_client_id();
+		$client_id    = WPCOM_OAuth::wpcom_client_id();
 		$redirect_uri = admin_url() . 'admin.php?page=' . $this->slug;
-		$access_token = self::get_access_token();
+		$access_token = WPCOM_OAuth::get_access_token();
 		$access_token = $access_token ? $access_token : '';
 		if ( is_wp_error( $access_token ) ) {
 			$access_token = '';
@@ -378,21 +259,12 @@ class Support_Wizard extends Wizard {
 	}
 
 	/**
-	 * Return client id of WPCOM auth app.
-	 *
-	 * @return string client id.
-	 */
-	public static function wpcom_client_id() {
-		return ( defined( 'NEWSPACK_WPCOM_CLIENT_ID' ) && NEWSPACK_WPCOM_CLIENT_ID ) ? NEWSPACK_WPCOM_CLIENT_ID : false;
-	}
-
-	/**
 	 * Check if wizard is configured and should be displayed.
 	 *
 	 * @return bool True if necessary variables are present.
 	 */
 	public static function configured() {
-		return self::support_api_url() && self::support_email() && self::wpcom_client_id();
+		return self::support_api_url() && self::support_email() && WPCOM_OAuth::wpcom_client_id();
 	}
 
 	/**
